requirements.txt
fastapi
uvicorn
google-genai
pytesseract
pdf2image
pymupdf
pillow
reportlab
jinja2